Extract debug code into its own partial
